+++ packages/components/src/components/Link/stories/Default.stories.tsx
@@ -10,15 +10,8 @@ const meta: Meta<typeof Link> = {
   component: Link,
   args: {
     onPress: action("onPress"),
-    variant: "default",
   },
   render: (props) => <Link {...props}>Adjust project</Link>,
-  argTypes: {
-    variant: {
-      control: "inline-radio",
-      options: ["default", "danger"],
-    },
-  },
   parameters: {
     controls: { exclude: ["onPress"] },
   },
@@ -38,11 +31,6 @@ export const Inline: Story = {
 };
 
 export const Disabled: Story = { args: { isDisabled: true } };
-
-export const Danger: Story = {
-  args: { variant: "danger" },
-  render: (props) => <Link {...props}>Delete project</Link>,
-};
 
 export const WithIcon: Story = {
   render: (props) => (