+++ packages/components/src/components/List/stories/Default.stories.tsx
@@ -12,7 +12,7 @@ import {
 } from "@/components/List";
 import type { AsyncDataLoader } from "@/components/List/model/loading/types";
 import { Avatar } from "@/components/Avatar";
-import { ContextMenu, ContextMenuItem } from "@/components/ContextMenu";
+import { ContextMenu, MenuItem } from "@/components/ContextMenu";
 import { Link } from "@/components/Link";
 import { IconDomain, IconSubdomain } from "@/components/Icon/components/icons";
 import StatusBadge from "@/components/StatusBadge";
@@ -74,8 +74,8 @@ const meta: Meta<typeof List> = {
               )}
 
               <ContextMenu>
-                <ContextMenuItem>Show details</ContextMenuItem>
-                <ContextMenuItem>Delete</ContextMenuItem>
+                <MenuItem>Show details</MenuItem>
+                <MenuItem>Delete</MenuItem>
               </ContextMenu>
             </>
           )}
@@ -107,8 +107,8 @@ export const ItemsWithLink: Story = {
               <Heading>{domain.hostname}</Heading>
               <Text>{domain.type}</Text>
               <ContextMenu>
-                <ContextMenuItem>Show details</ContextMenuItem>
-                <ContextMenuItem>Delete</ContextMenuItem>
+                <MenuItem>Show details</MenuItem>
+                <MenuItem>Delete</MenuItem>
               </ContextMenu>
             </Link>
           )}