+++ packages/docs/src/content/02-components/status/inline-alert/examples/with-content.tsx
@@ -1,6 +1,7 @@
 import InlineAlert from "@mittwald/flow-react-components/InlineAlert";
 import Heading from "@mittwald/flow-react-components/Heading";
 import Content from "@mittwald/flow-react-components/Content";
+import Button from "@mittwald/flow-react-components/Button";
 
 <InlineAlert>
   <Heading>Email address has been archived</Heading>
@@ -8,5 +9,6 @@ import Content from "@mittwald/flow-react-components/Content";
     As your domain has been deleted, this email address has
     been archived. To be able to send and receive emails,
     you must rename the address.
+    <Button>Update email address</Button>
   </Content>
 </InlineAlert>;