+++ packages/design-tokens/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.67](https://github.com/mittwald/flow/compare/0.1.0-alpha.66...0.1.0-alpha.67) (2024-04-16)
+
+### Features
+
+- add IllustratedMessage Component
+  ([25c5e97](https://github.com/mittwald/flow/commit/25c5e974c081a0d9bcbf0634a07d8df6cd9da91d))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
 # [0.1.0-alpha.65](https://github.com/mittwald/flow/compare/0.1.0-alpha.64...0.1.0-alpha.65) (2024-04-11)
 
 **Note:** Version bump only for package @mittwald/flow-design-tokens