+++ packages/stylesheet/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.67](https://github.com/mittwald/flow/compare/0.1.0-alpha.66...0.1.0-alpha.67) (2024-04-16)
+
+**Note:** Version bump only for package @mittwald/flow-stylesheet
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
 # [0.1.0-alpha.66](https://github.com/mittwald/flow/compare/0.1.0-alpha.65...0.1.0-alpha.66) (2024-04-12)
 
 **Note:** Version bump only for package @mittwald/flow-stylesheet